moved Plane to Potts, changed point type to Voxel, removed unused function
